feat: rename entries and calendar page
